pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "seeragents"
dynamic = ["version"]
description = "Seer - Multi-Agent System for Evaluating AI Agents"
readme = "README.md"
authors = [
  { name = "Lokesh Danu", email = "lokesh@getseer.dev" },
  { name = "Akshay Sharma", email = "akshay@getseer.dev" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.2",
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langgraph-api==0.5.33",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "langgraph-checkpoint>=3.0.1,<4.0.0",
    "langgraph-checkpoint-postgres>=3.0.1,<4.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "e2b==2.4.0",
    "e2b-code-interpreter==2.2.0",
    "tavily-python>=0.7.12",
    "langchain-chroma>=0.1.0",
    "tenacity>=9.1.2",
    "langchain-mcp-adapters==0.1.12",
    "aiosqlite>=0.20.0",
    "coolname==2.2.0",
    # CLI dependencies
    "click>=8.1.0",
    "rich>=13.0.0",
    "asyncpg>=0.31.0",
    "mlflow>=3.7.0",
    "tortoise-orm[asyncpg]>=0.25.2",
    "authlib>=1.6.6",
]

[tool.setuptools]
packages = [
  "api",
  "cli",
  "shared",
  "agents",
]
[tool.setuptools_scm]
fallback_version = "0.0.0"
#just for versioning

[tool.setuptools.package-data]
"*" = ["**/*.json", "**/*.txt"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

[project.optional-dependencies]
cli = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]


[project.scripts]
seer = "cli.main:main"


[project.urls]
Homepage = "https://www.getseer.dev"
Documentation = "https://github.com/seer-engg/seer#readme"
Repository = "https://github.com/seer-engg/seer"
Issues = "https://github.com/seer-engg/seer/issues"
